Split the API client into per-domain modules
lib/api.ts had grown to ~600 lines. Extract the fetch helpers (api/http.ts)
and shared types (api/types.ts), and split the api object into per-domain
modules (system, settings, stats, repos, packages, vulnerabilities,
relationships, rules, users). lib/api.ts now just composes them into the
single `api` object and re-exports the types — so every
`import { api, SomeType } from "@/lib/api"` is unchanged. Pure reorg; no
behavior change (full typecheck + build pass).
