+++ ui/mira/src/lib/api/http.ts
@@ -1,0 +1,54 @@
+// Shared fetch helpers for the API client. Domain modules under `lib/api/`
+// import these; `lib/api.ts` composes the modules into the single `api` object.
+
+// Empty default → same-origin requests (production single-container).
+// In dev, set VITE_API_URL=http://localhost:8100 in ui/mira/.env.local.
+export const API_BASE = import.meta.env.VITE_API_URL || ""
+
+export async function fetchJson<T>(path: string): Promise<T> {
+  const res = await fetch(`${API_BASE}${path}`, { credentials: "include" })
+  if (!res.ok) {
+    if (res.status === 401) window.location.href = "/login"
+    throw new Error(`API error ${res.status}: ${await res.text()}`)
+  }
+  return res.json() as Promise<T>
+}
+
+export async function postJson<T>(path: string, body: unknown): Promise<T> {
+  const res = await fetch(`${API_BASE}${path}`, {
+    method: "POST",
+    headers: { "Content-Type": "application/json" },
+    credentials: "include",
+    body: JSON.stringify(body),
+  })
+  if (!res.ok) throw new Error(`API error ${res.status}: ${await res.text()}`)
+  return res.json() as Promise<T>
+}
+
+export async function putJson<T>(path: string, body: unknown): Promise<T> {
+  const res = await fetch(`${API_BASE}${path}`, {
+    method: "PUT",
+    headers: { "Content-Type": "application/json" },
+    credentials: "include",
+    body: JSON.stringify(body),
+  })
+  if (!res.ok) throw new Error(`API error ${res.status}: ${await res.text()}`)
+  return res.json() as Promise<T>
+}
+
+export async function deleteJson(path: string): Promise<void> {
+  const res = await fetch(`${API_BASE}${path}`, {
+    method: "DELETE",
+    credentials: "include",
+  })
+  if (!res.ok) throw new Error(`API error ${res.status}: ${await res.text()}`)
+}
+
+export async function patchJson<T>(path: string): Promise<T> {
+  const res = await fetch(`${API_BASE}${path}`, {
+    method: "PATCH",
+    credentials: "include",
+  })
+  if (!res.ok) throw new Error(`API error ${res.status}: ${await res.text()}`)
+  return res.json() as Promise<T>
+}
